fix: per-generation capacity calculation and cursor pagination for dust queries (#997)
* fix: per-generation capacity calculation, cursor pagination, and i64 overflow for dust queries

* rename generation_query to generations_query

+++ indexer-api/src/infra/api/v4/subscription/dust_nullifier_transactions.rs
@@ -87,7 +87,7 @@ where
                 .map_err_into_client_error(|| "invalid hex-encoded nullifier prefix")?;
 
             let from = from_block.unwrap_or(0);
-            let to = to_block.unwrap_or(u64::MAX);
+            let to = to_block.unwrap_or(i64::MAX as u64);
 
             debug!("streaming existing dust nullifier transactions");
